requirements.txt
torch>=2.2.2
torch-geometric>=2.4.0
torchmetrics>=1.3.1
pytorch-lightning>=2.2.0
stable-baselines3[extra]>=2.2.0
gymnasium>=0.29.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
pyyaml>=6.0.1
osmnx>=1.9.3
sumolib>=1.19.0
traci>=1.19.0
networkx>=3.2.1
shapely>=2.0.1
geopandas>=0.14.4
matplotlib>=3.8.0
streamlit>=1.36.0
uvicorn>=0.29.0
fastapi>=0.111.0

